requirements.txt
pyspark==3.5.4
apache-sedona==1.6.1
geopandas
psycopg2-binary
jupyter
ipykernel